kisan-os/backend/requirements.txt
# KISAN-OS Backend - Complete Production Dependencies

# ============ CORE FRAMEWORK ============
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ============ DATABASE & ORM ============
sqlalchemy==2.0.24
sqlmodel==0.0.14
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
GeoAlchemy2==0.14.1
shapely==2.0.2

# ============ CACHING & MESSAGE QUEUE ============
redis==5.0.1
aioredis==2.0.1
celery==5.3.4

# ============ AUTHENTICATION & SECURITY ============
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7
jwt==1.3.1

# ============ HTTP & API CLIENTS ============
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# ============ VOICE & LANGUAGE PROCESSING ============
google-cloud-speech==2.20.1
google-cloud-text-to-speech==2.14.1
pyttsx3==2.90
wave==0.0.2

# ============ AI/ML & DEEP LEARNING ============
torch==2.1.2
torchvision==0.16.2
tensorflow==2.15.0
scikit-learn==1.3.2

# ============ DATA PROCESSING ============
pandas==2.1.4
numpy==1.26.4
pillow==10.1.0
opencv-python==4.8.1.78
Pillow-SIMD==9.2.0

# ============ VECTOR SEARCH & RAG ============
langchain==0.1.4
langchain-community==0.0.10
pinecone-client==3.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# ============ OPTIMIZATION (Logistics) ============
google-ortools==9.7.2996
scipy==1.11.4

# ============ WEB SCRAPING ============
beautifulsoup4==4.12.2
lxml==4.9.3

# ============ DATA SCIENCE & ANALYSIS ============
matplotlib==3.8.2
seaborn==0.13.1
scikit-image==0.22.0

# ============ LOGGING & MONITORING ============
python-json-logger==2.0.7
loguru==0.7.2

# ============ TESTING ============
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx[cli]==0.25.2

# ============ UTILITIES ============
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
python-multipart==0.0.6

